engine: fix resolving instance attachments in edit initialization mode
This is a bit more involved than anticipated in 88ee1b91c1f68d53ce9ba551bab334852e1e60cd (and a vindication of deferring that functionality until we could exercise it with real functionality!). This addresses the following gaps in that earlier variation:

- Binary data is read from the response in a way that works across all target runtimes (see previous commit)
- An attachment’s data type is determined by first looking at the `Blob` produced by a `Response`, then by inspecting the response `headers`, and then finally with a generic fallback when no content type can be determined.
